+++ packages/polaris-viz-native/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Removed `dataOffsetLeft` and `dataOffsetRight` props from `SparkBarChart` and unified them under the `TargetLine` prop.
 
 ## [6.5.0] - 2022-08-03